PseudoPotentialData

Enables programmatic access to standard pseudopotential libraries for solid-state calculations. In using this library the combination of a string identifier and the element symbol provides a unique and reproducible mapping to a pseudopotential file. Moreover in case the pseudopotential file happens to be missing on the computer Julia's artifact system takes care to automatically download it as needed.

Basic usage

For example, the following code automatically downloads the pseudopotential file of the stringent pseudodojo pseudopotential for LDA pseudopotentials (referred to by the identifier pd_nc_sr_lda_stringent_0.4.1_upf) and places the full path to the downloaded pseudopotential file into the filename variable:

using PseudoPotentialData
identifier = "pd_nc_sr_lda_stringent_0.4.1_upf"
family = PseudoFamily(identifier)
filename = pseudofile(family, :Si)

"/home/runner/.julia/artifacts/d7fdc17d088c5c7e87a7a107bec744a33e4338e2/Si.upf"

As you see this will be a string such as /home/user/.julia/artifacts/56094b8162385233890d523c827ba06e07566079/Si.upf, which luckily you don't usually have to know or remember. Note, further that this path may differ between computers, julia versions etc. It is therefore highly recommended to use the above mechanism based on identfier and element symbol instead of hard-coding the expanded path in user scripts.

For multiple elements you can similarly use

pseudofile.(family, [:C, :Si])

2-element Vector{String}:
 "/home/runner/.julia/artifacts/d7fdc17d088c5c7e87a7a107bec744a33e4338e2/C.upf"
 "/home/runner/.julia/artifacts/d7fdc17d088c5c7e87a7a107bec744a33e4338e2/Si.upf"

A PseudoFamily is furthermore an AbstractDict{Symbol,String} for the mapping of element symbol to file path, e.g. one can perform index lookup

family[:Si]

"/home/runner/.julia/artifacts/d7fdc17d088c5c7e87a7a107bec744a33e4338e2/Si.upf"

or iterate over pairs

for (k, v) in family
   println(k, " => ", v)
   break
end

Ag => /home/runner/.julia/artifacts/d7fdc17d088c5c7e87a7a107bec744a33e4338e2/Ag.upf

For a list of available identifiers see

PseudoPotentialData.family_identifiers()

20-element Vector{String}:
 "pd_nc_sr_pbesol_standard_0.4.1_psp8"
 "pd_nc_sr_pbe_standard_0.4.1_psp8"
 "pd_nc_sr_lda_standard_0.4.1_psp8"
 "pd_nc_sr_lda_stringent_0.4.1_upf"
 "pd_nc_fr_pbe_standard_0.4_psp8"
 "pd_paw_pbe_standard_1.1_xml"
 "pd_nc_sr_pbe_stringent_0.4.1_psp8"
 "pd_nc_sr_lda_standard_0.4.1_upf"
 "pd_nc_fr_pbesol_stringent_0.4_psp8"
 "pd_nc_fr_pbesol_stringent_0.4_upf"
 "pd_nc_sr_pbe_stringent_0.4.1_upf"
 "pd_nc_sr_pbesol_stringent_0.4.1_psp8"
 "pd_nc_fr_pbesol_standard_0.4_upf"
 "pd_nc_sr_pbesol_standard_0.4.1_upf"
 "pd_nc_fr_pbe_stringent_0.4_psp8"
 "pd_nc_fr_pbesol_standard_0.4_psp8"
 "pd_nc_sr_pbe_standard_0.4.1_upf"
 "pd_nc_fr_pbe_standard_0.4_upf"
 "pd_nc_sr_lda_stringent_0.4.1_psp8"
 "pd_nc_fr_pbe_stringent_0.4_upf"

More details on the meaning of these keys is given in the README of the PseudoLibrary repository.

Rename of pseudopotential identifier reserved

The current identifiers for the pseudopotential families is planned to be overhauled. This will be a breaking change, where the minor version of the package will be bumped.

Interface

PseudoPotentialData.families — Constant

The list of all known pseudopotential families.

PseudoPotentialData.PseudoFamily — Method

Construction of a PseudoFamily from a identifier representing the pseudopotential family to use. For a list of valid identifier, see family_identifiers. A PseudoFamily is an `AbstractDict{Symbol,String} mapping from an element symbol to the full path of the pseudopotential file.

PseudoPotentialData.artifact_directory — Method

Return the directory containing the pseudo files. This downloads the artifact if necessary.

PseudoPotentialData.available_elements — Method

Return the list of all pseudopotential files in the artifact

PseudoPotentialData.family_identifiers — Method

Get the list of available pseudopotential family identifiers.

PseudoPotentialData.pseudofile — Method

Get the full path to the file containing the pseudopotential information for a particular element (identified by an atomic symbol) and a particular pseudopotential family.