PseudoLibraries

Enables programmatic access to standard pseudopotential libraries in solid-state calculations. In using this library the combination of a string identifier and the element symbol provides a unique and reproducible mapping to a pseudopotential file. Moreover in case the pseudopotential file happens to be missing on the computer Julia's artifact system takes care to automatically download it as needed.

Basic usage

For example, the following code automatically downloads the pseudopotential file of the stringent pseudodojo pseudopotential for LDA pseudopotentials (referred to by the identifier pd_nc_sr_lda_stringent_0.4.1_upf) and places the full path to the downloaded pseudopotential file into the filename variable:

using PseudoLibraries
identifier = "pd_nc_sr_lda_stringent_0.4.1_upf"
library = PseudoLibrary(identifier)
filename = library[:Si]

"/home/runner/.julia/artifacts/d7fdc17d088c5c7e87a7a107bec744a33e4338e2/Si.upf"

As you see this will be a string such as /home/user/.julia/artifacts/56094b8162385233890d523c827ba06e07566079/Si.upf, which luckily you don't usually have to know or remember. Note, further that this path may differ between computers, julia versions etc. It is therefore highly recommended to use the above mechanism based on identfier and element symbol instead of hard-coding the expanded path in user scripts.

An alternative version to achieve the same thing as above is

filename = pseudofile(library, :Si)

"/home/runner/.julia/artifacts/d7fdc17d088c5c7e87a7a107bec744a33e4338e2/Si.upf"

or for multiple elements:

pseudofile.(library, [:C, :Si])

2-element Vector{String}:
 "/home/runner/.julia/artifacts/d7fdc17d088c5c7e87a7a107bec744a33e4338e2/C.upf"
 "/home/runner/.julia/artifacts/d7fdc17d088c5c7e87a7a107bec744a33e4338e2/Si.upf"

Some metadata information is stored in the library object:

library

PseudoLibrary("pd_nc_sr_lda_stringent_0.4.1_upf", "upf", "lda", v"0.4.1")

For a list of available identifiers see

PseudoLibraries.available_identifiers()

32-element Vector{String}:
 "pd_nc_sr_pbesol_standard_0.4.1_psp8"
 "pd_nc_sr_pbe_standard_0.4.1_psp8"
 "pd_nc_sr_lda_standard_0.4.1_psp8"
 "gbrv_pbesol_1.5_upf"
 "sssp_pbe_precision_1.1.2_upf"
 "sssp_pbesol_efficiency_1.1.2_upf"
 "pd_nc_sr_lda_stringent_0.4.1_upf"
 "pd_nc_fr_pbe_standard_0.4_psp8"
 "hgh_pbe_hgh"
 "pd_paw_pbe_standard_1.1_xml"
 ⋮
 "pd_nc_sr_pbesol_standard_0.4.1_upf"
 "pd_nc_fr_pbe_stringent_0.4_psp8"
 "pd_nc_fr_pbesol_standard_0.4_psp8"
 "pd_nc_sr_pbe_standard_0.4.1_upf"
 "pd_nc_fr_pbe_standard_0.4_upf"
 "sssp_pbe_efficiency_1.1.2_upf"
 "pd_nc_sr_lda_stringent_0.4.1_psp8"
 "pd_nc_fr_pbe_stringent_0.4_upf"
 "hgh_lda_upf"

More details on the meaning of these keys is given at PseudoLibrary.

Rename of keys reserved

The current system of identifying the pseudopotential families will change in the future. This will be a breaking change, where the minor version of the package will be bumped.

Interface

PseudoLibraries.PseudoLibrary — Method

Construction of a PseudoLibrary from a identifier representing the pseudopotential library to use.

PseudoLibraries.artifact_directory — Method

Return the directory containing the pseudo files. This downloads the artifact if necessary.

PseudoLibraries.available_identifiers — Method

Get the list of available pseudopotential library keys.

PseudoLibraries.pseudofile — Method

Get the full path to the file containing the pseudopotential information for a particular element and a particular pseudopotential library.